refactor: separate out collection view component
